requirements.txt
discord.py
python-dotenv
aiomysql
motor==3.3.2
pymongo==4.6.1
aiohttp>=3.8.0
opencv-python-headless
yt-dlp==2023.3.4
PyNaCl==1.5.0  # Required for voice support
